Update npm to 0.2.12-1
Also, other changes to the formula:

1. Prefer a config file rather than changing the hard-coded
defaultConfig module with a regexp.

2. Install modules in the typical node module path.  Though this still
requires that the user add it to their NODE_PATH environ manually, since
node builds up its require path based on execPath, at least it's just
one thing.  Also, npm will warn about this at runtime.  Installing
modules somewhere other than lib/node leads to a lot of people giving
incorrect advice to confused newcomers.  It'd be better to avoid that.

3. Install bins to the normal "bin" folder so that `npm install express`
and `npm update` will behave as expected.  Changing this is fine, but
it's important that npm's executable be installed in this folder as
well, or else `npm update npm` won't actually do that.

Also, apparently there's some magic where if it's named "share_bin", it
doesn't chmod the file 0755, but if the method is named "bin" it does.
I didn't dig into it.

4. Use the shasum in addition to md5.

Suggestions for further improvement:

It'd be good to install the npm lib folder and npm.js to lib/node as
well.  I couldn't work out how to do that.

Setting the configs and having npm install itself would be ideal.  In
fact, `installing npm` is probably the one thing that npm is best at.
I suspect that this recipe could be just a few lines, and a whole lot
easier to support.

However, every way that I could think to run `node cli.js install`
or `make` exited with an error and no output.  The exception to this
rule was to use backticks, which appeared to run the command in some
temp folder, fail to find some required files, and promptly exit.